feat(Stabilisation): Stabilisation

+++ back/src/service/UserService.js
@@ -17,14 +17,18 @@ class UserService extends BaseService {
                     if (res === true) {
                         const generatedToken = AccessGranted.generateToken(result.entityData.id);
                         this.services.github.updateClientUrl(result.entityData.github_token);
-
                         resolve({ user: DatastoreUtil.getEntityInformation(result), token: generatedToken });
                     } else {
-                        reject(new Error(MessageUtil.getErrors().INVALID_CREDENTIAL.fr));
+                        reject(new Error(MessageUtil.getErrors().INVALID_CREDENTIAL.fr).message);
                     }
-                    resolve(res === true);
                 }).catch(err => this.rejectAndLogError(reject, err.message));
-            }).catch(err => this.rejectAndLogError(reject, err.message));
+            }).catch((err) => {
+                if (err.message === 'User not found') {
+                    reject(new Error(MessageUtil.getErrors().INVALID_CREDENTIAL.fr).message);
+                } else {
+                    this.rejectAndLogError(reject, err.message);
+                }
+            });
         });
     }
 
@@ -40,12 +44,42 @@ class UserService extends BaseService {
         });
     }
 
-    updateUser(user) {
+    updateUser(session, user) {
         return new Promise((resolve, reject) => {
-            user.password = undefined;
-            this.dao.update(user)
-                .then(result => resolve(DatastoreUtil.getEntityInformation(result)))
-                .catch(err => this.rejectAndLogError(reject, err.message));
+            if (session.user.id !== user.id) {
+                reject(new Error(MessageUtil.getErrors().UPDATE_OTHER_USER.fr));
+            } else {
+                user.password = undefined;
+                this.dao.update(user)
+                    .then(result => resolve(DatastoreUtil.getEntityInformation(result)))
+                    .catch(err => this.rejectAndLogError(reject, err.message));
+            }
+        });
+    }
+
+    updatePassword(session, userId, oldUserPassword, userPassword, userPasswordConfirm) {
+        return new Promise((resolve, reject) => {
+            if (userPasswordConfirm === userPassword) {
+                this.dao.getById(userId).then((result) => {
+                    BcryptUtil.validPassword(result.entityData.password, oldUserPassword).then((res) => {
+                        if (res === true) {
+                            if (session.user.id !== userId) {
+                                this.rejectAndLogError(reject, new Error(MessageUtil.getErrors().UPDATE_OTHER_USER.fr).message);
+                            } else {
+                                BcryptUtil.generatePassword(userPassword).then((hash) => {
+                                    this.dao.update({ id: userId, password: hash })
+                                        .then(updatedUser => resolve(DatastoreUtil.getEntityInformation(updatedUser)))
+                                        .catch(err => this.rejectAndLogError(reject, err.message));
+                                }).catch(err => this.rejectAndLogError(reject, err.message));
+                            }
+                        } else {
+                            reject(new Error(MessageUtil.getErrors().INVALID_CREDENTIAL.fr));
+                        }
+                    }).catch(err => this.rejectAndLogError(reject, err.message));
+                }).catch(err => this.rejectAndLogError(reject, err.message));
+            } else {
+                this.rejectAndLogError(reject, new Error(MessageUtil.getErrors().NOT_SAME_PASSWORD.fr).message);
+            }
         });
     }
 
@@ -61,11 +95,23 @@ class UserService extends BaseService {
         });
     }
 
-    deleteUser(userId) {
+    deleteUser(session, userId) {
         return new Promise((resolve, reject) => {
-            this.dao.deleteById(userId)
-                .then(user => resolve(user))
-                .catch(err => this.rejectAndLogError(reject, err.message));
+            if (session.user.id !== userId) {
+                reject(new Error(MessageUtil.getErrors().DELETE_OTHER_USER.fr));
+            } else {
+                this.getUser(userId).then((user) => {
+                    if (user !== null) {
+                        this.services.repository.deleteAllByOwner(userId).then(() => {
+                            this.dao.deleteById(userId)
+                                .then(result => resolve(result))
+                                .catch(err => this.rejectAndLogError(reject, err.message));
+                        }).catch(err => this.rejectAndLogError(reject, err.message));
+                    } else {
+                        reject(new Error(MessageUtil.getErrors().USER_NOT_FOUND.fr));
+                    }
+                }).catch(err => this.rejectAndLogError(reject, err.message));
+            }
         });
     }
 
@@ -73,14 +119,6 @@ class UserService extends BaseService {
         return new Promise((resolve, reject) => {
             this.dao.getById(userId)
                 .then(result => resolve(DatastoreUtil.getEntityInformation(result)))
-                .catch(err => this.rejectAndLogError(reject, err.message));
-        });
-    }
-
-    findUsers() {
-        return new Promise((resolve, reject) => {
-            this.dao.findAll()
-                .then(result => resolve(result))
                 .catch(err => this.rejectAndLogError(reject, err.message));
         });
     }
